Issue #116: Develop Os Command Extension
* Refactored all unit tests
* Refactored Super Connector Integration Test
* Added Javadoc
* Code enhancement
