+++ metricshub-oscommand-extension/src/it/resources/os/SuperConnectorOsIT/connectors/Hardware/Hardware.yaml
