+++ metricshub-oscommand-extension/src/it/resources/os/SuperConnectorOsIT/connectors/SuperConnectorCommon-header/embeddedFile-22
