+++ metricshub-oscommand-extension/src/it/resources/os/SuperConnectorOsIT/expected/expected.json
