+++ metricshub-oscommand-extension/src/main/java/org/sentrysoftware/metricshub/extension/oscommand/OsCommandRequestExecutor.java
@@ -35,6 +35,9 @@ import org.sentrysoftware.metricshub.engine.common.exception.ClientException;
 import org.sentrysoftware.metricshub.engine.common.helpers.LoggingHelper;
 import org.sentrysoftware.ssh.SshClient;
 
+/**
+ * Provides functionality to execute remote SSH commands, manage SSH authentication, and transfer files to remote hosts.
+ */
 @Slf4j
 public class OsCommandRequestExecutor {
